// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  type: 'watercolour' | 'oil' | 'ceramics';
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface ISimpleContentBlock {
  title?: string | null;
  img?: (number | null) | Media;
  content?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  layout?: ('img-content' | 'content-img') | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'SimpleContentBlock';
}